requirements.txt
django==3.2.18
django-truncate==0.1
psycopg2-binary==2.9.5
bitstring==4.0.1
base58==2.1.1
coinaddrvalidator==1.1.3
requests==2.28.1
fire
icon_font_to_png
django-health-check==3.17.0
gunicorn==20.1.0
Pillow==9.4.0
django.storages==1.13.2
boto3==1.26.60
django-s3-storage==0.13.11
numpy==1.24.1
schedule==1.1.0
django-rest-swagger==2.2.0
django-aws-secrets-manager==0.1.5
django-recaptcha==4.0.0
#coinaddrvalidator==1.1.3
watchtower==3.0.1
django-simple-captcha==0.5.20
django_ratelimit==4.1.0
django-ajax-selects==2.2.0
django-allauth==0.58.2
qrcode==7.1
aws-xray-sdk==2.12.1
numerize==0.12
matplotlib==3.8.2
pandas==2.1.3
inflection==0.5.1
yfinance
